Move utility functions to ovsp4rt_utils
- Extracted utility functions from ovsp4rt.cc and moved them to
  ovsp4rt_utils.cc.

- Defined prototypes for the functions in ovsp4rt_utils.h.
