+++ ovs-p4rt/sidecar/CMakeLists.txt
@@ -15,6 +15,8 @@ add_library(ovs_sidecar_o OBJECT
     ovsp4rt_credentials.h
     ovsp4rt_session.cc
     ovsp4rt_session.h
+    ovsp4rt_utils.cc
+    ovsp4rt_utils.h
 )
 
 if(DPDK_TARGET)
